некоторые изменения

+++ BlogNetworkB/Controllers/RoleController.cs
@@ -8,20 +8,27 @@ using ConnectionLib.DAL.Queries;
 using System.Security.Claims;
 using BlogNetworkB.BLL.Models.Role;
 using ConnectionLib.DAL.Queries.Role;
+using Microsoft.AspNetCore.Mvc.Filters;
+using BlogNetworkB.Infrastructure.Exceptions;
+using BlogNetworkB.Models;
+using BlogNetworkB.Models.CustomError;
 
 namespace BlogNetworkB.Controllers
 {
+    [ExceptionHandler]
     [Route("/Roles")]
     public class RoleController : Controller
     {
         readonly IRoleRepository _roleRepository;
         readonly IAuthorRepository _authorRepository;
         readonly IMapper _mapper;
-        public RoleController(IRoleRepository roleRepository, IAuthorRepository authorRepository, IMapper mapper)
+        readonly ILogger<RoleController> _logger;
+        public RoleController(IRoleRepository roleRepository, IAuthorRepository authorRepository, IMapper mapper, ILogger<RoleController> logger)
         {
             _roleRepository = roleRepository;
             _authorRepository = authorRepository;
             _mapper = mapper;
+            _logger = logger;
         }
 
         #region получение списка ролей
@@ -55,6 +62,8 @@ namespace BlogNetworkB.Controllers
 
             MyRoleViewModel mrvm = new() { Roles = myRoleList.ToArray() };
 
+            _logger.LogInformation("Пользователь {email} обратился к списку своих ролей", HttpContext.User.Claims.FirstOrDefault().Value);
+
             return View(mrvm);
         }
 
@@ -72,27 +81,45 @@ namespace BlogNetworkB.Controllers
         [Route("/[controller]/AuthorRoles/NewRole")]
         public async Task<IActionResult> ConfirmAddRole([FromForm] RoleViewModel roleViewModel)
         {
+            var admin = HttpContext.User.Claims.FirstOrDefault().Value;
             var role = await _roleRepository.GetRoleById(roleViewModel.RoleId);
-            var author = await _authorRepository.GetAuthorById(roleViewModel.AuthorId);
 
-            // т.к. метод get отказывается принимать
-            // roleViewModel.AuthorId
-            if (role == null)
+            try
             {
-                int id = roleViewModel.AuthorId;
-                return View("AddRole", id);
-            }
+                // т.к. метод get отказывается принимать
+                // roleViewModel.AuthorId
+                if (role == null)
+                {
+                    //ModelState.AddModelError("RoleId", "роль с таким id отсутствует");
 
-            // если автор уже имеет такую роль - ничего не происходит
-            bool hasRole = (_authorRepository.GetAuthorsRoles(author).Result.Contains(role) == true);
-            if (hasRole)
-            {
+                    throw new CustomException($"Пользователь {admin} пытался добавить роль с id={roleViewModel.RoleId}, такой роли нет");
+
+                    //return View("AddRole", roleViewModel);
+                }
+
+                var author = await _authorRepository.GetAuthorById(roleViewModel.AuthorId);
+
+                // если автор уже имеет такую роль - ничего не происходит
+                bool hasRole = (_authorRepository.GetAuthorsRoles(author).Result.Contains(role) == true);
+                if (hasRole)
+                {
+                    //ModelState.AddModelError("RoleId", "Пользователь уже имеет такую роль");
+
+                    throw new CustomException($"Пользователь {admin} пытался добавить пользователю {author.Email} роль с id={roleViewModel.RoleId} повторно");
+                }
+
+                await _authorRepository.AddRole(author, role);
+
+                _logger.LogInformation("Пользователь {admin} добавил пользователю {author} роль {id}", admin, author.Email, roleViewModel.RoleId);
+
                 return RedirectToAction("AuthorsList", "Author");
             }
-
-            await _authorRepository.AddRole(author, role);
-
-            return RedirectToAction("AuthorsList", "Author");
+            catch (Exception ex)
+            {
+                _logger.LogError("{error}", ex.Message);
+                CustomErrorViewModel cevm = new() { Message = ex.Message };
+                return View("/Views/Alert/SomethingWrong.cshtml", cevm);
+            }
         }
 
         [Authorize]
@@ -108,24 +135,42 @@ namespace BlogNetworkB.Controllers
             var role = await _roleRepository.GetRoleById(rvm.RoleId);
             var author = await _authorRepository.GetAuthorById(rvm.AuthorId);
 
-            // т.к. метод get отказывается принимать
-            // roleViewModel.AuthorId
-            if (role == null)
+            try
             {
-                int id = rvm.AuthorId;
-                return View("RemoveRole", id);
-            }
+                // т.к. метод get отказывается принимать
+                // roleViewModel.AuthorId
+                if (role == null)
+                {
+                    throw new CustomException($"Роли с id={rvm.RoleId} не существует");
+                }
 
-            // если автор не имеет такую роль
-            bool hasRole = (_authorRepository.GetAuthorsRoles(author).Result.Contains(role) != true);
-            if (hasRole)
+                var authorRoles = await _authorRepository.GetAuthorsRoles(author);
+
+                // если автор не имеет такую роль
+                bool hasRole = authorRoles.Contains(role) == true;
+
+                if (!hasRole)
+                {
+                    throw new CustomException($"Попытка удаления отсутвующей роли у пользователя");
+                }
+
+                if (hasRole && role.Name == "User")
+                {
+                    throw new CustomException("Попытка удаления базовой роли");
+                }
+
+                await _authorRepository.DeleteAuthorRole(author, role);
+
+                _logger.LogInformation("Пользователь {admin} удалил у пользователя {email} роль {id}", HttpContext.User.Claims.FirstOrDefault().Value, author.Email, role.Id);
+
+                return RedirectToAction("AuthorsList", "Author");
+            }
+            catch (Exception ex)
             {
-                return View("/Views/Alert/SomethingWrong.cshtml");
+                _logger.LogError("{error}", ex.Message);
+                CustomErrorViewModel cevm = new CustomErrorViewModel { Message = ex.Message };
+                return View("/Views/Alert/SomethingWrong.cshtml", cevm);
             }
-
-            await _authorRepository.DeleteAuthorRole(author, role);
-
-            return RedirectToAction("AuthorsList", "Author");
         }
 
         #endregion
@@ -146,15 +191,28 @@ namespace BlogNetworkB.Controllers
 
                 try
                 {
+                    var sameRole = _roleRepository.GetAll().Result.Any(r => r.Name == role.Name);
+
+                    if (sameRole != default)
+                    {
+                        throw new CustomException($"Попытка добавления дубликата. Роль с именем \'{role.Name}\' уже существует");
+                    }
+
                     await _roleRepository.AddRole(role);
+
+                    _logger.LogInformation("Пользователь {email} добавил роль {roleName}", HttpContext.User.Claims.FirstOrDefault().Value, role.Name);
 
                     return RedirectToAction("RoleList");
                 }
-                catch
+                catch(Exception ex)
                 {
-                    return View("/Views/Alert/SomethingWrong.cshtml");
+                    _logger.LogError("{error}", ex.Message);
+                    CustomErrorViewModel cevm = new() { Message = ex.Message };
+                    return View("/Views/Alert/SomethingWrong.cshtml", cevm);
                 }
             }
+
+            _logger.LogWarning("Неверно заполнена форма");
 
             return View("AddRoleToRoles");
         }
@@ -182,10 +240,14 @@ namespace BlogNetworkB.Controllers
 
                 await _roleRepository.UpdateRole(role, _mapper.Map<UpdateRoleQuery>(request));
 
+                _logger.LogInformation("Пользователь {email} изменил описание у роли {id}", HttpContext.User.Claims.FirstOrDefault().Value, role.Id);
+
                 return RedirectToAction("RoleList");
             }
 
             int id = urrvm.RoleId;
+
+            _logger.LogWarning("Неверно заполнена форма");
 
             return RedirectToAction("EditRoleDescription", id);
         }
@@ -196,8 +258,25 @@ namespace BlogNetworkB.Controllers
         public async Task<IActionResult> DeleteRole(int id)
         {
             var role = await _roleRepository.GetRoleById(id);
-            await _roleRepository.DeleteRole(role);
-            return RedirectToAction("RoleList");
+            try
+            {
+                if (role.Name == "User" || role.Name == "Moderator" || role.Name == "Admin")
+                {
+                    throw new CustomException($"Попытка удалить одну из стандартных ролей: {role.Name}");
+                }
+
+                await _roleRepository.DeleteRole(role);
+
+                _logger.LogInformation("Пользователь {email} удалил роль с {id}", HttpContext.User.Claims.FirstOrDefault().Value, role.Id);
+
+                return RedirectToAction("RoleList");
+            }
+            catch(Exception ex)
+            {
+                _logger.LogError("{error}", ex.Message);
+                CustomErrorViewModel cevm = new() { Message = ex.Message };
+                return View("/Views/Alert/SomethingWrong.cshtml", cevm);
+            }
         }
     }
 }
